+++ examples/11_fs-router/src/pages/foo/index.tsx
@@ -7,4 +7,10 @@ const Foo = () => (
   </div>
 );
 
+export const getConfig = () => {
+  return {
+    render: 'dynamic',
+  } as const;
+};
+
 export default Foo;